agent/requirements.txt
fastapi
uvicorn[standard]
pydantic-ai-slim[openrouter]
websockets
python-dotenv